requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
enum-tools==0.12.0
ffmpeg==1.4
frozenlist==1.5.0
h11==0.14.0
idna==3.10
isodate==0.7.2
lxml==5.3.1
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
streamlink==7.1.3
trio==0.29.0
trio-websocket==0.12.0
twitchAPI==4.4.0
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
